dseq_search_window is now of type uint8_t again, it is unpacked to 16 bit for computing the dprofile

+++ src/algo/8/search_8_util.h
@@ -61,7 +61,7 @@ struct s8info {
 };
 
 static inline uint8_t move_db_sequence_window_8( uint8_t c, uint8_t * d_begin[CHANNELS_8_BIT],
-        uint8_t * d_end[CHANNELS_8_BIT], uint16_t dseq_search_window[CHANNELS_8_BIT * CDEPTH_8_BIT] ) {
+        uint8_t * d_end[CHANNELS_8_BIT], uint8_t dseq_search_window[CHANNELS_8_BIT * CDEPTH_8_BIT] ) {
     for( int j = 0; j < CDEPTH_8_BIT; j++ ) {
         if( d_begin[c] < d_end[c] ) {
             dseq_search_window[CHANNELS_8_BIT * j + c] = *(d_begin[c]++);
@@ -79,8 +79,8 @@ static inline uint8_t move_db_sequence_window_8( uint8_t c, uint8_t * d_begin[CH
 p_s8info search_8_sse41_init( p_search_data sdp );
 p_s8info search_8_avx2_init( p_search_data sdp );
 
-void dprofile_fill_8_sse41( __mxxxi * dprofile, uint16_t * dseq_search_window );
-void dprofile_fill_8_avx2( __mxxxi * dprofile, uint16_t * dseq_search_window );
+void dprofile_fill_8_sse41( __mxxxi * dprofile, uint8_t * dseq_search_window );
+void dprofile_fill_8_avx2( __mxxxi * dprofile, uint8_t * dseq_search_window );
 
 void search_8_sse41_sw( p_s8info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, uint8_t query_id );
 void search_8_sse41_nw( p_s8info s, p_db_chunk chunk, p_minheap heap, p_node * overflow_list, uint8_t query_id );
